+++ e2e/tests-dfx/assetscanister.bash
@@ -34,6 +34,16 @@ check_permission_failure() {
     fi
 }
 
+@test "dfx deploy does not retry forever on permissions failure" {
+  install_asset assetscanister
+  dfx_start
+  assert_command dfx deploy
+  echo "new file content" > 'src/e2e_project_frontend/assets/new_file.txt'
+
+  assert_command_fail dfx deploy --identity anonymous
+  assert_contains "Caller does not have Prepare permission"
+}
+
 @test "validation methods" {
   assert_command dfx identity new controller --storage-mode plaintext
   assert_command dfx identity use controller